이벤트 글 삭제, 수정, 댓글 권한부여

+++ src/main/java/com/example/marumaru_sparta_verspring/service/MeetService.java
@@ -43,7 +43,6 @@ public class MeetService {
         meet.setContent(meetRequestDto.getContent());
         meet.setAddress(meetRequestDto.getAddress());
         meet.setDate(meetRequestDto.getDate());
-
         meetRepository.save(meet);
         return meet;
     }
@@ -77,35 +76,56 @@ public class MeetService {
     }
 
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, Long userId) {
         Meet meet = meetRepository.findById(id)
                 .orElseThrow(() ->
                         new IllegalArgumentException("해당 게시글이 없습니다."));
-        meetRepository.delete(meet);
+        if (meet.getUserId() != userId) {
+            throw new IllegalArgumentException("작성자만 삭제 가능");
+        } else {
+            meetRepository.delete(meet);
+        }
     }
 
     @Transactional
-    public Meet update(Long id, MeetUpdateRequestDto meetUpdateRequestDto) {
+    public Meet update(Long id, MeetUpdateRequestDto meetUpdateRequestDto, Long userId) {
         Meet meet = meetRepository.findById(id)
                 .orElseThrow(() ->
                         new IllegalArgumentException("해당 게시글이 없습니다."));
-        meet.setTitle(meetUpdateRequestDto.getTitle());
-        meet.setContent(meetUpdateRequestDto.getContent());
-        meetRepository.save(meet);
-        return meet;
+
+        if (meet.getUserId() != userId) {
+            throw new IllegalArgumentException("작성자만 수정 가능");
+        } else {
+            meet.setTitle(meetUpdateRequestDto.getTitle());
+            meet.setContent(meetUpdateRequestDto.getContent());
+            meetRepository.save(meet);
+            return meet;
+        }
     }
 
     @Transactional
-    public void deleteComment(Long id) {
-        meetCommentRepository.deleteById(id);
+    public void deleteComment(Long id, Long userId) {
+        MeetComment meetComment = meetCommentRepository.findById(id)
+                .orElseThrow(() ->
+                        new IllegalArgumentException("해당 게시글이 없습니다."));
+        if (meetComment.getUser().getId() != userId) {
+            throw new IllegalArgumentException("작성자만 삭제 가능");
+        } else {
+            meetCommentRepository.deleteById(id);
+        }
     }
 
     @Transactional
-    public void updateComment(MeetCommentRequestDto meetCommentRequestDto) {
+    public void updateComment(MeetCommentRequestDto meetCommentRequestDto, Long userId) {
         MeetComment meetComment = meetCommentRepository.findById(meetCommentRequestDto.getIdx()).orElseThrow(
                 () -> new NullPointerException("해당 아이디가 존재하지 않습니다.")
         );
-        meetComment.setComment(meetCommentRequestDto.getComment());
-        meetCommentRepository.save(meetComment);
+
+        if (meetComment.getUser().getId() != userId) {
+            throw new IllegalArgumentException("작성자만 수정 가능");
+        } else {
+            meetComment.setComment(meetCommentRequestDto.getComment());
+            meetCommentRepository.save(meetComment);
+        }
     }
 }
